# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance backend API for HiDocs Form \u0026 Exam Maker application using Golang and Domain-Driven Design (DDD).",
        "title": "HiDocs Backend API (Form \u0026 Exam Maker)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "HiDocs Support",
            "email": "support@hidocs.id"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/forms/{form_id}/analytics": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get real-time analytics for charts (Pie Chart score distribution, Question accuracy bar charts)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "form_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.FormAnalytics"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/forms/{form_id}/export": {
            "get": {
                "tags": [
                    "Responses"
                ],
                "summary": "Export form responses to Excel or CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "form_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Export format ('xlsx' or 'csv', default: 'xlsx')",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/forms/{form_id}/responses": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Responses"
                ],
                "summary": "Get all responses for a form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "form_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.ResponseDetailDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/forms/{form_id}/submit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Responses"
                ],
                "summary": "Submit form / exam answers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "form_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Submit Form Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SubmitFormRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SubmitResponseResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/responses/{response_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Responses"
                ],
                "summary": "Get individual response details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Response ID",
                        "name": "response_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ResponseDetailDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/responses/{response_id}/grade": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Responses"
                ],
                "summary": "Adjust score/grade manually for a response",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Response ID",
                        "name": "response_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Grade Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GradeResponseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "domain.FormAnalytics": {
            "type": "object",
            "properties": {
                "average_score": {
                    "type": "number"
                },
                "highest_score": {
                    "type": "number"
                },
                "lowest_score": {
                    "type": "number"
                },
                "question_breakdown": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/domain.QuestionAnalytics"
                    }
                },
                "total_responses": {
                    "type": "integer"
                }
            }
        },
...
        "domain.QuestionAnalytics": {
            "type": "object",
            "properties": {
                "accuracy_rate": {
                    "type": "number"
                },
                "correct_count": {
                    "type": "integer"
                },
                "option_counts": {
                    "description": "Option ID -\u003e count",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "question_id": {
                    "type": "string"
                },
                "question_text": {
                    "type": "string"
                },
                "total_answered": {
                    "type": "integer"
                }
            }
        },
...
        "dto.AnswerDetailDTO": {
            "type": "object",
            "properties": {
                "answer_text": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "points_earned": {
                    "type": "number"
                },
                "question_id": {
                    "type": "string"
                },
                "question_text": {
                    "type": "string"
                },
                "score_given": {
                    "type": "number"
                },
                "selected_option_id": {
                    "type": "string"
                },
                "selected_option_text": {
                    "type": "string"
                }
            }
        },
...
        "dto.GradeResponseRequest": {
            "type": "object",
            "properties": {
                "total_score": {
                    "type": "number",
                    "minimum": 0
                }
            }
        },
...
        "dto.ResponseDetailDTO": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AnswerDetailDTO"
                    }
                },
                "form_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_auto_submitted": {
                    "type": "boolean"
                },
                "respondent_email": {
                    "type": "string"
                },
                "submitted_at": {
                    "type": "string"
                },
                "total_score": {
                    "type": "number"
                }
            }
        },
        "dto.SubmitAnswerDetail": {
            "type": "object",
            "required": [
                "question_id"
            ],
            "properties": {
                "answer_text": {
                    "type": "string"
                },
                "question_id": {
                    "type": "string"
                },
                "selected_option_id": {
                    "type": "string"
                }
            }
        },
        "dto.SubmitFormRequest": {
            "type": "object",
            "required": [
                "answers",
                "respondent_email"
            ],
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SubmitAnswerDetail"
                    }
                },
                "is_auto_submitted": {
                    "type": "boolean"
                },
                "passcode": {
                    "type": "string"
                },
                "respondent_email": {
                    "type": "string"
                }
            }
        },
        "dto.SubmitResponseResult": {
            "type": "object",
            "properties": {
                "is_auto_submitted": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "response_id": {
                    "type": "string"
                },
                "submitted_at": {
                    "type": "string"
                },
                "total_score": {
                    "type": "number"
                }
            }
        },
...
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}